+++ gerenciamento/src/main/java/Servlet/servlet_cadastro_e_atualizacao_produtos.java
@@ -103,7 +103,7 @@ public class servlet_cadastro_e_atualizacao_produtos extends servlet_recuperacao
 
 				String id = request.getParameter("id");
 
-				ModelProdutos dadosJsonUser1 = daoproduto.consultaProduto(Integer.parseInt(id), super.getUsuarioLogado(request).getId());
+				ModelProdutos dadosJsonUser1 = daoproduto.consultaProduto(Long.parseLong(id), super.getUsuarioLogado(request).getId());
 				List<ModelFornecimento> dadosJsonUser2 = daoFornecimento.listarFornecedores(Integer.parseInt(id));
 				Gson gson = new Gson();
 				String json1 = gson.toJson(dadosJsonUser1);
@@ -116,7 +116,7 @@ public class servlet_cadastro_e_atualizacao_produtos extends servlet_recuperacao
 
 				String id = request.getParameter("id");
 
-				ModelProdutos dadosJsonUser = daoproduto.consultaProduto(Integer.parseInt(id), super.getUsuarioLogado(request).getId());
+				ModelProdutos dadosJsonUser = daoproduto.consultaProduto(Long.parseLong(id), super.getUsuarioLogado(request).getId());
 				Gson gson = new Gson();
 				String json = gson.toJson(dadosJsonUser);
 				PrintWriter printWriter = response.getWriter();
